Add files via upload
The Titanic dataset project involves creating a predictive model to determine whether passengers aboard the Titanic survived or not. It is a popular beginner's project due to the accessibility of the dataset. The dataset contains various details about each passenger, including age, gender, ticket class, fare, cabin, and survival status. By utilizing this data, the goal is to develop a machine learning model that can make accurate predictions about a passenger's survival outcome based on these attributes. This project serves as a practical introduction to data analysis, feature engineering, and machine learning, offering insights into the tragic historical event of the Titanic sinking.